Add support for sending non-sync values (#13)
* Add a manual implementation of Sync for SPSCQueue

This allows the spsc queue to be used to send consumers and producers
of other spsc queues, or more generally any types that are `Send` but
not `Sync`.

Without this change the following occurs:

- The queue's value type `T` includes an `SPSCConsumer<_>`,
  which is `!Sync`.
- The non-sync `T` is embedded in `SPSCQueue` as `PhantomData<T>`,
  which makes it also `!Sync`.
- `Arc<T>` is only `Send` when `T: Send + Sync`.
- So then `SPSCConsumer` and `SPSCProducer` (which both have
  `Arc<SPSCQueue<T, B>>` fields) become `!Send`.

* Add Sync impls to remaining queue types to allow sending `!Sync` values

Following on from the previous commit, the `mpsc` and `spmc` queues
also require manual `Sync` implementations to allow them to send values
that are `Send` but `!Sync`.